requirements.txt
glfw==2.1.0
numpy==1.20.1
opencv-contrib-python==4.5.1.48
Pillow==8.1.2
pycairo==1.20.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
paho-mqtt==1.5.1
